Added medalgin helper file
